requirements.txt
APScheduler==3.9.1.post1
Django==4.2
django_allauth==0.52.0
django_apscheduler==0.6.2
django_filter==22.1
celery==5.4.0
environs==11.0.0